backend/requirements.txt
flask
flask-cors
numpy
requests
gunicorn
Pillow
mediapipe==0.10.14
